// src/services/models/memePaginationResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { MemeResponse } from "./memeResponse";

export interface MemePaginationResponse {
  readonly total: number;
  readonly page: number;
  readonly pageSize: number;
  readonly data: readonly MemeResponse[];
}